requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
pandas
sqlalchemy
asyncpg
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
cachetools
nest-asyncio
slowapi
python-multipart
# LlamaIndex + integrations
llama-index-core
llama-index-embeddings-openai
llama-index-llms-groq
llama-index-llms-openai
llama-index-vector-stores-postgres
llama-cloud-services
# Optional tracing / evals
langsmith
